Create STAXParser
One of my first new files of code. We needed to go through an XML file and find the location code as quickly as possible. This is what I came up with.

Challenges included learning STAX, learning/using streams, overloading to handle strings or Lists of strings.

I thought it went pretty well, comments or thoughts for optimization always welcomed.